requirements.txt
python-telegram-bot
numpy
Pillow
tensorflow-cpu
opencv-python-headless
scikit-image
